feat: :sparkles: add modal-not-identified-payment and add related styles
