+++ src/modules/clients/components/payments-tab/modal-identify-payment-action/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./modal-identify-payment-action";
