--- src/modules/clients/components/payments-tab/modal-identify-payment/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./modal-identify-payment";
