+++ src/modules/clients/components/payments-tab/modal-not-identified-payment/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./modal-not-identified-payment";
